Modifications to Entity and Parser, and created World.
Parser is still not returning Entities or Maps, will change with Alpha 0.2.
World is now functionally a container of Entities. Entities should not have
a position unless they are in the world. In the future, world might change to
hold on to the positions of its static entities on its own to save on memory
usage, but for now it relies on the Entity itself to keep track of its position.