Update for the parser, now it handles xml files.
Sadly, it doesn't really do anything.. but atleast it understands the
xml now. And the editor still understands the tileset, so yay.
Read last entry for more information about the example.xml.

Also, added a new file, mapexample.xml, this is just for testing
purposes, nothing will have to be changed in the maps resulting
from the map editor.